requirements.txt
google-cloud-vision
picamera
gpiozero
mpyg321

# sudo apt-get install espeak
# sudo apt-get install espeak python-espeak

